% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  port = as.numeric(Sys.getenv("RESVIEW_PORT")),
  launch.browser = TRUE,
  ...
)
}
\arguments{
\item{...}{A series of options to be used inside the app.}
}
\description{
Run the Shiny Application
}
